# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TravelPlan API
  version: 1.0.0
  description: >-
    # TravelPlan Public API


    Production-grade AI travel management API.


    ## Versioning


    - **Current:** `/api/v1`

    - **Legacy (compatible):** `/api` (same handlers; receives `X-API-Version:
    1`)


    Prefer `/api/v1` for all new integrations.


    ## Authentication


    Most endpoints require a JWT Bearer token from `POST /api/v1/auth/login` or
    signup.


    ```

    Authorization: Bearer <token>

    ```


    Admin monitoring APIs additionally require an admin role (`User.role=admin`
    or `ADMIN_EMAILS`).


    ## Rate limits


    Redis-backed limits (fail-open if Redis is down):


    | Scope | Limit |

    |-------|-------|

    | Login | 10 / minute |

    | Signup | 10 / minute |

    | Forgot password | 5 / minute |

    | AI endpoints | 20 / hour |

    | Public weather | 60 / minute |


    Responses may include `X-RateLimit-*` headers. HTTP **429** when exceeded.


    ## Errors


    Error bodies use:


    ```json

    { "success": false, "code": "ERROR_CODE", "message": "Human readable",
    "error": "..." }

    ```
  contact:
    name: TravelPlan API
    email: hello@travelplan.com
  license:
    name: Proprietary
    identifier: LicenseRef-TravelPlan-Proprietary
servers:
  - url: http://localhost:5000
    description: Local development (paths include /api/v1)
  - url: https://api.travelplan.example
    description: Production host (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT from /auth/login or /auth/signup. Send as `Authorization:
        Bearer <token>`."
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data: {}
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
        error:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          example: TRIP_CREATED
        category:
          type: string
        title:
          type: string
        message:
          type: string
        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        status:
          type: string
          enum:
            - UNREAD
            - READ
            - ARCHIVED
        read:
          type: boolean
        actionUrl:
          type: string
        trip:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/notifications:
    get:
      summary: List notifications
      description: List notifications
      tags: &a154
        - Notifications
      parameters: &a155 []
      responses: &a156
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
      operationId: listNotifications
      security: *a24
  /api/notifications:
    get:
      summary: List notifications
      description: |-
        List notifications

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a154
      parameters: *a155
      responses: *a156
      operationId: listNotifications
      security: *a24
      x-deprecated-prefer: Use /api/v1listNotifications
...
  /api/v1/notifications/{id}:
    delete:
      summary: Delete notification
      description: Delete notification
      tags: &a166
        - Notifications
      parameters: &a167
        - name: id
          in: path
          required: true
          description: id identifier
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses: &a168
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: deleteNotification
      security: *a24
  /api/notifications/{id}:
    delete:
      summary: Delete notification
      description: |-
        Delete notification

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a166
      parameters: *a167
      responses: *a168
      operationId: deleteNotification
      security: *a24
      x-deprecated-prefer: Use /api/v1deleteNotification
  /api/v1/notifications/settings:
    get:
      summary: Notification settings
      description: Notification settings
      tags: &a169
        - Notifications
      parameters: &a170 []
      responses: &a171
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: getNotificationSettings
      security: *a24
    put:
      summary: Update notification settings
      description: Update notification settings
      tags: &a172
        - Notifications
      parameters: &a173 []
      responses: &a174
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: updateNotificationSettings
      security: *a24
  /api/notifications/settings:
    get:
      summary: Notification settings
      description: |-
        Notification settings

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a169
      parameters: *a170
      responses: *a171
      operationId: getNotificationSettings
      security: *a24
      x-deprecated-prefer: Use /api/v1getNotificationSettings
    put:
      summary: Update notification settings
      description: |-
        Update notification settings

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a172
      parameters: *a173
      responses: *a174
      operationId: updateNotificationSettings
      security: *a24
      x-deprecated-prefer: Use /api/v1updateNotificationSettings
...